Restrict unconfirmed users to create new user accounts.
