+++ app/__init__.py
@@ -10,6 +10,7 @@ bootstrap = Bootstrap()
 mail = Mail()
 moment = Moment()
 db = SQLAlchemy()
+
 login_manager = LoginManager()
 login_manager.login_view = 'auth.login'
 
@@ -25,12 +26,13 @@ def create_app(config_name):
     db.init_app(app)
     login_manager.init_app(app)
 
-    # Pievienot saites un pielāgotas lapas ar kļūdas paziņojumiem
-
     from .main import main as main_blueprint
     app.register_blueprint(main_blueprint)
 
     from .auth import auth as auth_blueprint
     app.register_blueprint(auth_blueprint, url_prefix = '/auth')
+    
+    from .reg import reg as reg_blueprint
+    app.register_blueprint(reg_blueprint, url_prefix = '/reg')
 
     return app
