+++ app/auth/forms.py
@@ -1,7 +1,6 @@
 from flask_wtf import FlaskForm
-from flask import current_app
 from wtforms import StringField, PasswordField, BooleanField, SubmitField, ValidationError
-from wtforms.validators import DataRequired, Length, Email, Regexp, EqualTo
+from wtforms.validators import DataRequired, Length, Email, EqualTo
 from ..models import User
 
 
@@ -12,27 +11,6 @@ class LoginForm(FlaskForm):
     remember_me = BooleanField('Atcerēties mani')
     sumbit = SubmitField('Pieslēgties')
 
-
-class RegistrationForm(FlaskForm):
-    email = StringField(
-        'E-pasts', validators=[DataRequired(), Length(1, 64), Email()])
-    username = StringField('Lietotāja vārds', validators=[DataRequired(), Length(1, 64), Regexp(
-        '^[A-Za-z][A-Za-z0-9_.]*$', 0, 'Usernames must have only letters, numbers, dots or ''underscores')])
-    # password = PasswordField('Parole', validators=[
-    #                          DataRequired(), EqualTo('password2', message='Paroles nesakrīt')])
-    # password2 = PasswordField('Apstiprini paroli', validators=[DataRequired()])
-    # Hidden because new user registration is done by admin and password is generated
-    submit = SubmitField('Reģistrēt')
-
-    def validate_email(self, field):
-        app = current_app._get_current_object()
-        if User.query.filter_by(email=field.data).first():
-            raise ValidationError('Šāds e-pasts jau ir reģistrēts')
-        email_domain = field.data.split('@')[-1]
-
-    def validate_username(self, field):
-        if User.query.filter_by(username=field.data).first():
-            raise ValidationError('Šāds lietotāja vārds jau ir reģistrēts')
 
 
 class ChangePasswordForm(FlaskForm):
