+++ app/auth/views.py
@@ -4,10 +4,24 @@ from . import auth
 from .. import db
 from ..models import User
 from ..email import send_email
-from .forms import LoginForm, RegistrationForm, ChangePasswordForm, PasswordResetRequestForm, PasswordResetForm, ChangeEmailForm
-from ..decorators import admin_required, permission_required
-import string
-import secrets
+from .forms import LoginForm, ChangePasswordForm, PasswordResetRequestForm, PasswordResetForm, ChangeEmailForm
+
+
+@auth.before_app_request
+def before_request():
+    if current_user.is_authenticated \
+            and not current_user.confirmed \
+            and request.endpoint \
+            and request.blueprint != 'auth' \
+            and request.endpoint != 'static':
+        return redirect(url_for('auth.unconfirmed'))
+
+
+@auth.route('/unconfirmed')
+def unconfirmed():
+    if current_user.is_anonymous or current_user.confirmed:
+        return redirect(url_for('main.index'))
+    return render_template('auth/unconfirmed.html')
 
 
 @auth.route('/login', methods=['GET', 'POST'])
@@ -25,31 +39,12 @@ def login():
     return render_template('auth/login.html', form=form)
 
 
-@auth.route('/logut')
+@auth.route('/logout')
 @login_required
 def logout():
     logout_user()
     flash('Jūs esat atslēdzies no sistēmas')
     return redirect(url_for('main.index'))
-
-
-@auth.route('/register', methods=['GET', 'POST'])
-@login_required
-@admin_required
-def register():
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        symbols = string.ascii_letters + string.digits
-        user = User(email=form.email.data.lower(),
-                    username=form.username.data, password=''.join(secrets.choice(symbols) for i in range(16)))
-        db.session.add(user)
-        db.session.commit()
-        token = user.generate_confirmation_token()
-        send_email(user.email, 'Apstiprini savu lietotāja kontu',
-                   'auth/email/confirm', user=user, token=token)
-        flash('Uz Jūsu e-pasta adresi nosūtīts aicinājums apstiprināt to!')
-        return redirect(url_for('main.index'))
-    return render_template('auth/register.html', form=form)
 
 
 @auth.route('/confirm/<token>')
@@ -64,22 +59,6 @@ def confirm(token):
         flash(
             'Saite e-pasta adreses apstiprināšanai ir nederīga vai tai ir beidzies termiņš')
     return redirect(url_for('main.index'))
-
-
-@auth.before_app_request
-def before_request():
-    if current_user.is_authenticated \
-            and not current_user.confirmed \
-            and request.blueprint != 'auth' \
-            and request.endpoint != 'static':
-        return redirect(url_for('auth.unconfirmed'))
-
-
-@auth.route('/unconfirmed')
-def unconfirmed():
-    if current_user.is_anonymous or current_user.confirmed:
-        return redirect(url_for('main.index'))
-    return render_template('auth/unconfirmed.html')
 
 
 @auth.route('/confirm')
