+++ app/reg/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+reg = Blueprint('reg', __name__)
+
+from . import views
