+++ app/templates/base.html
@@ -32,7 +32,7 @@
                         <li><a href="{{ url_for('auth.change_password') }}">Nomainīt paroli</a></li>
                         <li><a href="{{ url_for('auth.change_email_request') }}">Nomainīt e-pasta adresi</a></li>
                         {% if current_user.is_administrator() %}
-                        <li><a href="{{ url_for('auth.register') }}">Reģistrēt jaunu lietotāju</a></li>
+                        <li><a href="{{ url_for('reg.register') }}">Reģistrēt jaunu lietotāju</a></li>
                         {% endif %}                    
                         <li><a href="{{ url_for('auth.logout') }}">Atvienoties</a></li>
                     </ul>
